--- app/backend/tests/pytest.ini
@@ -1,5 +1,0 @@
-[pytest]
-testpaths = tests
-python_files = test_*.py
-python_functions = test_*
-addopts = -v --cov=src
